+++ python/core_universal/applitools/core_universal/__init__.py
@@ -1,6 +1,6 @@
 from __future__ import absolute_import
 
-__version__ = "4.0.1"
+__version__ = "2.3.5"
 
 
 def get_instance():