+++ python/core_universal/applitools/core_universal/server.py
@@ -6,16 +6,22 @@ from pkg_resources import resource_filename
 
 logger = getLogger(__name__)
 
-_exe_name = "eyes-universal.exe" if sys.platform == "win32" else "eyes-universal"
-executable_path = resource_filename("applitools.eyes_universal", "bin/" + _exe_name)
+_exe_name = "core.exe" if sys.platform == "win32" else "core"
+executable_path = resource_filename("applitools.core_universal", "bin/" + _exe_name)
 
 
 class SDKServer(object):
     log_file_name = None  # backward compatibility with eyes-selenium<=5.6
 
     def __init__(self):
-        """Start eyes-universal service subprocess and obtain its port number."""
-        command = [executable_path, "--no-singleton", "--shutdown-mode", "stdin"]
+        """Start core-universal service subprocess and obtain its port number."""
+        command = [
+            executable_path,
+            "universal",
+            "--no-singleton",
+            "--shutdown-mode",
+            "stdin",
+        ]
         # Capture and keep stdin reference to notify USDK when it should terminate.
         # USDK is expected to terminate when it receives EOF on its stdin.
         # The pipe is automatically closed and EOF is sent by OS when python terminates.