+++ python/tests/unit/core_universal/test_instance.py
@@ -5,7 +5,7 @@ from time import sleep
 import psutil
 import pytest
 
-from applitools.eyes_universal import get_instance
+from applitools.core_universal import get_instance
 
 
 def test_server_instance():
@@ -15,12 +15,12 @@ def test_server_instance():
 
 
 @pytest.mark.parametrize("reason", ["normal", "crash"])
-def test_automatic_eyes_universal_termination_on_python_termination(reason):
-    # A simple python program that executes eyes-universal and waits on stdin
+def test_automatic_core_universal_termination_on_python_termination(reason):
+    # A simple python program that executes core-universal and waits on stdin
     code = (
         "import ctypes;"
         "import sys;"
-        "from applitools.eyes_universal import get_instance;"
+        "from applitools.core_universal import get_instance;"
         "print(get_instance());"
         "sys.stdout.flush();"
         "sys.stdin.readline();"
@@ -30,20 +30,20 @@ def test_automatic_eyes_universal_termination_on_python_termination(reason):
     python_process = Popen([sys.executable, "-c", code], stdin=PIPE, stdout=PIPE)
     output_line = python_process.stdout.readline()
     assert output_line.startswith(b"SDKServer")
-    (eyes_universal,) = psutil.Process(python_process.pid).children()
-    assert eyes_universal.is_running()
+    (core_universal,) = psutil.Process(python_process.pid).children()
+    assert core_universal.is_running()
     python_process.stdin.write(b"\n")
     python_process.stdin.flush()
     python_process.wait()
-    sleep(1)  # eyes-universal receives EOF on stdin and terminates, might take time
+    sleep(1)  # core-universal receives EOF on stdin and terminates, might take time
 
     # psutils returns True when asking .is_running() on zombie process
     # so have to check status explicitly
     try:
-        eyes_universal_status = eyes_universal.status()
+        core_universal_status = core_universal.status()
     except psutil.NoSuchProcess:
-        eyes_universal_status = "terminated"
+        core_universal_status = "terminated"
     # GitHub Actions running tests in alpine container do not reap
-    # terminated eyes_universal leaving it as a zombie.
+    # terminated core_universal leaving it as a zombie.
     # Looks like there is nothing we can do about it except accept.
-    assert eyes_universal_status in ("terminated", "zombie")
+    assert core_universal_status in ("terminated", "zombie")