+++ python/tests/unit/test_cyclic_imports.py
@@ -5,7 +5,7 @@ from sys import executable
 import pytest
 
 import EyesLibrary
-from applitools import common, core, eyes_universal, images, selenium
+from applitools import common, core, core_universal, images, selenium
 
 
 def enum_modules(package):
@@ -21,7 +21,7 @@ def enum_modules(package):
 
 
 @pytest.mark.parametrize(
-    "pkg", [common, core, images, selenium, eyes_universal, EyesLibrary]
+    "pkg", [common, core, images, selenium, core_universal, EyesLibrary]
 )
 def test_cyclic_imports(pkg):
     for module in enum_modules(pkg):